requirements.txt
langchain==0.3.21
langchain-anthropic==0.3.10
langchain-community==0.3.20
langchain-core==0.3.47
langchain-text-splitters==0.3.7
psycopg2-binary==2.9.10
python-dotenv==1.0.1